userdashboard

+++ src/main/webapp/WEB-INF/views/ListCategory.jsp
@@ -58,10 +58,9 @@
  												<td><%=cb.getCategoryId()%></td>
 												<td><%=cb.getCategoryName()%></td>
 												<td><%=cb.getDeleted()%></td>
-												<td><a class="btn btn-info">Edit</a> <a
-													class="btn btn-danger"
-													href="deletecategory/<%=cb.getCategoryId()%>"
-													onclick="return confirm('Are you sure want to delete this record?')">Delete</a>			
+												<td>
+												<a style="color:blue;"><i class="bi bi-pencil-square" style="font-size:20px;"></i></a>
+												<a href="deletecategory/<%=cb.getCategoryId()%>" onclick="return confirm('Are you sure want to delete this record?')" style="color:red;"><i class="bi bi-trash" style="font-size:20px;"></i></a></td>			
 											</tr>
 											<%
 											}
@@ -69,7 +68,7 @@
 										</tbody>
 									</table>
 								</div>
-                        <a href="newcategory"><div style="color:black","margin-left:1070px;" class="icon" ><i class="ri-add-box-fill fs-2"></i></div></a>
+                        <a href="newcategory" style="color:black;"><div style="margin-left:1070px;" class="icon" ><i class="ri-add-box-fill fs-2"></i></div></a>
 								<!-- End Table with stripped rows -->
 </div>
 							</div>
